Initial commit
This is zulip-to-gotify.py, which I wrote and tweaked over two days in
May 2019 and haven't touched since. I've had it running since then
without issue.

This is essentially the exact code I've been running; the only change
is `import sys` and `GOTIFY_POST_URL = sys.argv[1]` instead of the URL
being hardcoded.